app/requirements.txt
cassandra-driver
venv-pack
pyarrow
fastparquet
wheel
pathvalidate
pyspark
tqdm
cqlsh